Ignore setting empty path in wxGTK wxFileDialog.
wxGenericFileButton calls wxFileDialog::SetPath() with empty path if no
initial path is explicitly specified and this clearly shouldn't do anything --
but it did, as it set the file dialog path to the initial directory (cwd by
default), and so opened it in the parent of that directory which was
completely unexpected and not useful at all.

Simply don't do anything when SetPath("") is called.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@70752 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775