Split documentation of the GUI wxLog classes in a separate file.
Put the GUI classes in a new interface/wx/generic/logg.h to make it easier to
generate wrappers for only the base or only the core libraries.

Also change the order of the classes remaining in log.h to ensure that the
base classes are always defined before the derived ones.

Closes #14784.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@72818 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775