XRC spec: relax requirements on some commonly omitted properties.
Some properties documented as required, such as labels on wxStaticText
or wxCheckBox or the size of wxSizer spacers, are frequently omitted in
practice.

Rather than strictly requiring them, forcing the markup to include dummy
empty elements, loosen the restriction in the spec and make them
optional.  Nothing changes about actual acceptance of files by
wxXmlResource: it would already silently deal with the lack of these
properties.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@74924 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775