Fix syntax error in the dialogs sample after r75028.
An extra "&&" was left in the preprocessor check, remove it.

Also remove the check for __WXMAC_OSX__ as it's unneeded as
USE_NATIVE_FONT_DIALOG_FOR_MACOSX is only defined as 1 under wxOSX and,
actually, wrong, as __WXMAC_OSX__ is never defined anywhere and was probably a
corruption of __WXOSX_MAC__ in the first place.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@75033 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775